Added Parse* pipes in the controllers and modified the movement interceptor

+++ src/common/databases/postgresql/entities/movement.entity.ts
@@ -22,7 +22,7 @@ export class MovementEntity {
   @Column("character varying", { name: "mov_reason", length: 500 })
   reason: string;
 
-  @Column("bigint", { name: "mov_amount" })
+  @Column("integer", { name: "mov_amount" })
   amount: number;
 
   @Column("integer", { name: "mov_fees", default: () => "1" })
